- fix: preview missing on long filenames
- feat: add data annotations to org dto
- fix: camelcase serialisation breaking the signals
- feat: #1087 add warning when create an org
- fix: total length should be 3
- feat: #1089 allow org searching
- refactor: bad file extensions
- refactor: use a common authorize attribute
- feat: #1088 show org suggestions when creating org
- feat: display message on banner not errors
- feat: use azure search not the dumb filter
- fix: possible null references

Related work items: #1087, #1088, #1089

+++ apps/api/Services/Basic/BasicSearchService.cs
@@ -6,6 +6,7 @@ using Amphora.Api.Models.Search;
 using Amphora.Common.Contracts;
 using Amphora.Common.Models.Amphorae;
 using Amphora.Common.Models.DataRequests;
+using Amphora.Common.Models.Organisations;
 
 namespace Amphora.Api.Services.Basic
 {
@@ -13,12 +14,15 @@ namespace Amphora.Api.Services.Basic
     {
         private readonly IAmphoraeService amphoraeService;
         private readonly IEntityStore<DataRequestModel> dataRequestStore;
+        private readonly IEntityStore<OrganisationModel> organisationStore;
 
-        public BasicSearchService(
-            IAmphoraeService amphoraeService, IEntityStore<DataRequestModel> dataRequestStore)
+        public BasicSearchService(IAmphoraeService amphoraeService,
+                                  IEntityStore<DataRequestModel> dataRequestStore,
+                                  IEntityStore<OrganisationModel> organisationStore)
         {
             this.amphoraeService = amphoraeService;
             this.dataRequestStore = dataRequestStore;
+            this.organisationStore = organisationStore;
         }
 
         // this just returns all the amphora is memory
@@ -48,6 +52,8 @@ namespace Amphora.Api.Services.Basic
 
         public async Task<EntitySearchResult<T>> SearchAsync<T>(string searchText, SearchParameters parameters) where T : ISearchable
         {
+            searchText ??= ""; // ensure searchText not null
+
             if (typeof(T) == typeof(AmphoraModel))
             {
                 var entities = new List<AmphoraModel>();
@@ -56,7 +62,7 @@ namespace Amphora.Api.Services.Basic
                     a => a.Name.Contains(searchText) || a.Description.Contains(searchText));
                 entities.AddRange(res);
 
-                if (parameters.Skip.HasValue && parameters.Top.HasValue)
+                if (parameters != null && parameters.Skip.HasValue && parameters.Top.HasValue)
                 {
                     var toRemove = parameters.Skip * parameters.Top;
                     entities.RemoveRange(0, toRemove.Value);
@@ -73,13 +79,35 @@ namespace Amphora.Api.Services.Basic
                     a => a.Name.Contains(searchText) || a.Description.Contains(searchText));
                 entities.AddRange(res);
 
-                if (parameters.Skip.HasValue && parameters.Top.HasValue)
+                if (parameters != null && parameters.Skip.HasValue && parameters.Top.HasValue)
                 {
                     var toRemove = parameters.Skip * parameters.Top;
                     entities.RemoveRange(0, toRemove.Value);
                 }
 
                 var take = parameters.Top ?? entities.Count;
+                return new EntitySearchResult<T>(entities.Take(take).ToList().Cast<T>());
+            }
+            else if (typeof(T) == typeof(OrganisationModel))
+            {
+                var entities = new List<OrganisationModel>();
+
+                var res = await organisationStore.QueryAsync(
+                    a => a.Name.Contains(searchText) || a.About.Contains(searchText));
+                entities.AddRange(res);
+
+                if (parameters != null && parameters.Skip.HasValue && parameters.Top.HasValue)
+                {
+                    var toRemove = parameters.Skip * parameters.Top;
+                    entities.RemoveRange(0, toRemove.Value);
+                }
+
+                var take = 10;
+                if (parameters != null)
+                {
+                    take = parameters.Top ?? entities.Count;
+                }
+
                 return new EntitySearchResult<T>(entities.Take(take).ToList().Cast<T>());
             }
             else
